Make tests work for non-PouchDB
