+++ test/client/replicate.go
@@ -7,7 +7,6 @@ import (
 	"time"
 
 	"github.com/flimzy/kivik"
-	"github.com/flimzy/kivik/driver/pouchdb/bindings"
 	"github.com/flimzy/kivik/test/kt"
 )
 
@@ -30,19 +29,6 @@ func replicate(ctx *kt.Context) {
 func callReplicate(ctx *kt.Context, client *kivik.Client, target, source, repID string, opts kivik.Options) (*kivik.Replication, error) {
 	opts = replicationOptions(ctx, client, target, source, repID, opts)
 	return client.Replicate(context.Background(), target, source, opts)
-}
-
-func replicationOptions(ctx *kt.Context, client *kivik.Client, target, source, repID string, in map[string]interface{}) map[string]interface{} {
-	if in == nil {
-		in = make(map[string]interface{})
-	}
-	if ctx.String("mode") != "pouchdb" {
-		in["_id"] = repID
-		return in
-	}
-	in["source"] = bindings.GlobalPouchDB().New(source, ctx.Options("db"))
-	in["target"] = bindings.GlobalPouchDB().New(target, ctx.Options("db"))
-	return in
 }
 
 func testReplication(ctx *kt.Context, client *kivik.Client) {
